requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
qdrant-client==1.15.1
ollama==0.4.4
pydantic==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.9     
pypdf==4.3.1                
python-docx==1.1.2            